Refactor: separate rendering code for Mermaid and Graphviz
