+++ crates/solidity/outputs/cargo/tests/src/bindings_output/graph/mod.rs
@@ -1,0 +1,2 @@
+pub mod graphviz;
+pub mod mermaid;
